fix(connection): always promisify `connection.query`
After setting `connection.query = Util.promisify(connection.query)`,
`connection.query` still works with callbacks as well.

If you called `getDb()` or `getConnection()` with a callback it just
takes the connection as-is from the pool. If you called it with a
callback it mutated the original connection object to contain the
promisifed version of the `.query()` method. This caused the pool to
potentially contain both promisified and not promisified versions of the
method. If you would get a non-promisified version of the method and you
would call it with async await, it would return the Query object instead
of properly waiting for the result.